Response-Profile id is long

+++ tests/ovp/java2/src/test/java/com/kaltura/client/test/ResponseProfileTest.java
@@ -505,7 +505,7 @@ public class ResponseProfileTest extends BaseTest{
 		APIOkRequestsExecutor.getExecutor().queue(requestBuilder.build(client));
 	}
 
-	protected void deleteResponseProfile(int id) throws Exception {
+	protected void deleteResponseProfile(long id) throws Exception {
 		startAdminSession();
 		DeleteResponseProfileBuilder requestBuilder = ResponseProfileService.delete(id);
 		APIOkRequestsExecutor.getExecutor().queue(requestBuilder.build(client));
